Add grade and get Grade by id
Missing Actual grade, but there's a connection between a grade, a student and an exam.

+++ mini-project-2/MiniProject2.Factory/Clients/ExamClient.cs
@@ -8,7 +8,7 @@ namespace MiniProject2.Factory.Clients
 {
     public class ExamClient
     {
-        public static async Task<ExamDTO> GetExamByIdAsync(int id)
+        public static async Task<ExamDTO> GetExamByIdAsync(long id)
         {
             // The port number(5001) must match the port of the gRPC server.
             using var channel = GrpcChannel.ForAddress("http://serviceexam:80");
