+++ mini-project-2/MiniProject2.Factory/Factory/StudentFactory.cs
@@ -1,12 +1,13 @@
 ﻿using Google.Protobuf.WellKnownTypes;
 using MiniProject2.Factory.Clients;
 using MiniProject2.Models.DTO;
+using System;
 
 namespace MiniProject2.Factory{
 
   public class StudentFactory
   {
-    public static async Task<StudentDTO> getStudentById(int id)
+    public static async Task<StudentDTO> getStudentById(long id)
     {
       return await StudentClient.GetStudentByIdAsync(id);
     }
